backend/food_security/requirements.txt
Django==3.2
PyMySQL==1.0.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework_simplejwt==4.7.0